container/requirements.txt
devtools
django
django-cors-headers
django-ninja
uvicorn[standard]
whitenoise
